Create AND3 layout generator
The AND3 layout generator produces the following ports:
- A
- B
- C
- Y
- vss0
- vss1
- vdd0
- vdd1
- vdd2

A, B, and C are the inputs; Y is the output; vssN is ground; vddN is
power. Note that the power and ground ports are not internally
connected. Users of the gate must connect power/ground to ALL
appropriate ports, not just one of them.