Place decoder ntaps and ptaps
Since there isn't enough space to place taps between decoder gates,
this commit adds taps to the left and right sides of the decoders.

Some of the pin structure has been changed – in particular, merging of
decoder wells and power pins is done in the decoder generator, rather
than in the top-level SRAM bank generator.