Update rustfmt import granularity settings (#142)
Add a rustfmt.toml file specifying that the import granularity
should be at the module level. Note that this feature is currently
unstable, so you will need a nightly version of Rust to run rustfmt.

Once a nightly version of Rust is installed, run the formatter
by running `cargo +nightly fmt` in the `sram22/` directory.