Add v1 replica bitline control logic + testbench (#330)
This replica timing scheme discharges the bitlines more than necessary,
but it is compatible with SKY130's replica bitline cell structure,
which connects the replica wordline to the main memory cell wordline.

This is better than the inverter-chain based approach, which has the
problem of poor matching between the memory array and the delay chain.
It also eliminates the need to guess and check the correct inverter
delay chain length for each SRAM size.